auto_processor: split the unit tests into separate files.
